@@ -20,7 +20,7 @@ class Seeker {
     while (_running) {
       Duration newPosition = player.position + positionInterval;
       if (newPosition < Duration.zero) newPosition = Duration.zero;
-      if (newPosition > mediaItem.duration) newPosition = mediaItem.duration;
+      if (newPosition > mediaItem.duration!) newPosition = mediaItem.duration!;
       player.seek(newPosition);
       await Future.delayed(stepInterval);
     }
